changed float to double to make rendering more clear

+++ inc/intersections.h
@@ -6,7 +6,7 @@
 /*   By: jpancorb <jpancorb@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 16:18:10 by irozhkov          #+#    #+#             */
-/*   Updated: 2025/02/13 20:10:43 by jpancorb         ###   ########.fr       */
+/*   Updated: 2025/02/17 14:32:04 by irozhkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 # include "minirt.h"
 
 void	get_intersections(t_scene *scene, t_ray *ray);
-void	check_ray(t_ray *ray, int color, float dist, t_type_id id);
+void	check_ray(t_ray *ray, int color, double dist, t_type_id id);
 
 #endif
