+++ src/intersections/intersections.c
@@ -6,7 +6,7 @@
 /*   By: irozhkov <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 14:12:55 by irozhkov          #+#    #+#             */
-/*   Updated: 2025/02/12 15:27:20 by irozhkov         ###   ########.fr       */
+/*   Updated: 2025/02/17 14:32:20 by irozhkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ void	get_intersections(t_scene *scene, t_ray *ray)
 	}
 }
 
-void	check_ray(t_ray *ray, int color, float dist, t_type_id id)
+void	check_ray(t_ray *ray, int color, double dist, t_type_id id)
 {
 	if (ray->hit == 0)
 	{
@@ -52,7 +52,5 @@ void	check_ray(t_ray *ray, int color, float dist, t_type_id id)
 			else
 				ray->dot_color = color;
 		}
-//		else if (dist > ray->dist_curr)
-//			ray->is_shadow = 1;
 	}
 }
